refactor: Moved apollo client queries and schema to graphql
