+++ ui/src/graphql/schema.ts
@@ -1,0 +1,50 @@
+// TODO at some point, I'll integrate a tool to generate types from queries
+export type CollectionInfo = {
+  id: string;
+  name: string;
+  major_ver: number;
+  minor_ver: number;
+  is_locked: boolean;
+  created_by: {
+    email: string;
+  };
+  created_at: number;
+};
+
+export type User = {
+  email: string;
+  created_at?: number;
+  second_last_login_at?: number;
+  last_login_at?: number;
+};
+
+export type ColumnDef = {
+  header: string;
+  name: string;
+  name_en: string;
+  name_fr: string;
+  description_en: string;
+  description_fr: string;
+  data_type: string;
+  conditions: {
+    condition_type: string;
+    parameters: string[];
+  }[];
+};
+export type Collection = {
+  id: string;
+  stable_key?: string;
+  major_ver: number;
+  minor_ver: number;
+  is_current_version: boolean;
+  created_by: User;
+  created_at: number;
+  name_en: string;
+  name_fr: string;
+  description_en: string;
+  description_fr: string;
+  is_locked: boolean;
+  owners: User[];
+  uploaders: User[];
+  column_defs: ColumnDef[];
+};
