+++ ui/src/schema/utils.ts
@@ -65,41 +65,3 @@ export const constructErrorMessage = (error: ErrorObject): string => {
       return `${error.instancePath} does not satisfy the schema requirements.`;
   }
 };
-
-export interface User {
-  email: string;
-  created_at?: number;
-  second_last_login_at?: number;
-  last_login_at?: number;
-}
-
-export interface ColumnDef {
-  header: string;
-  name: string;
-  name_en: string;
-  name_fr: string;
-  description_en: string;
-  description_fr: string;
-  data_type: string;
-  conditions: {
-    condition_type: string;
-    parameters: string[];
-  }[];
-}
-export interface Collection {
-  id: string;
-  stable_key?: string;
-  major_ver: number;
-  minor_ver: number;
-  is_current_version: boolean;
-  created_by: User;
-  created_at: number;
-  name_en: string;
-  name_fr: string;
-  description_en: string;
-  description_fr: string;
-  is_locked: boolean;
-  owners: User[];
-  uploaders: User[];
-  column_defs: ColumnDef[];
-}
